+++ projects/demo/src/app/component-test/component-test.component.spec.ts
@@ -3,7 +3,6 @@ import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { provideHttpClient } from '@angular/common/http';
 import { provideHttpClientTesting } from '@angular/common/http/testing';
-import { NguiAutoCompleteModule } from 'auto-complete';
 
 import { ComponentTestComponent } from './component-test.component';
 import { AppService } from '../app.service';
@@ -14,7 +13,7 @@ describe('ComponentTestComponent', () => {
 
 	beforeEach(waitForAsync(() => {
 		TestBed.configureTestingModule({
-			imports: [FormsModule, NguiAutoCompleteModule, ComponentTestComponent],
+			imports: [FormsModule, ComponentTestComponent],
 			providers: [AppService, provideHttpClient(), provideHttpClientTesting()],
 			schemas: [NO_ERRORS_SCHEMA],
 		}).compileComponents();